requirements.txt
requests~=2.28.2
python-dotenv~=0.21.1
py-cord==2.4.0
pymongo~=4.3.3
pika~=1.3.1
pika-stubs~=0.1.3
py-consul~=1.2.4
